The magic bit.
Whatever warps our IP back by a byte also clobbers rax. But it doesn't
do so when the IP is at a syscall insn, for some reason. Maybe this
would all make more sense if I were able to find the relevant bit of
code, but I'm not even entirely sure it's in the kernel, though I think
it should be in sigreturn.